+++ packages/print/src/index.ts
@@ -6,6 +6,10 @@
  * @copyright 2019-2022 Nguyen Huu Phuoc <me@phuoc.ng>
  */
 
+export { getAllPagesNumbers } from './getAllPagesNumbers';
+export { getCustomPagesNumbers } from './getCustomPagesNumbers';
+export { getEvenPagesNumbers } from './getEvenPagesNumbers';
+export { getOddPagesNumbers } from './getOddPagesNumbers';
 export type { PrintProps } from './Print';
 export { PrintIcon } from './PrintIcon';
 export type { PrintMenuItemProps } from './PrintMenuItem';